+++ packages/module-service-discovery/src/provider.ts
@@ -27,8 +27,8 @@ export interface IServiceDiscoveryProvider {
     ): Promise<IHttpClient>;
 
     configureClient(
-        config: ModulesConfigurator<[HttpModule]>,
-        serviceName: string | { key: string; alias: string }
+        serviceName: string | { key: string; alias: string },
+        config: ModulesConfigurator<[HttpModule]>
     ): Promise<void>;
 }
 
@@ -62,8 +62,8 @@ export class ServiceDiscoveryProvider implements IServiceDiscoveryProvider {
     }
 
     public async configureClient(
-        config: ModulesConfigurator<[HttpModule]>,
-        serviceName: string | { key: string; alias: string }
+        serviceName: string | { key: string; alias: string },
+        config: ModulesConfigurator<[HttpModule]>
     ): Promise<void> {
         const { key, alias } =
             typeof serviceName === 'string'