+++ packages/module-services/src/configurator.ts
@@ -1,0 +1,8 @@
+import { ApiClientFactory } from './types';
+
+export interface IApiConfigurator {
+    /** Method for creating IHttpClients */
+    createClient?: ApiClientFactory;
+}
+
+export class ApiConfigurator implements IApiConfigurator {}