+++ packages/module-http/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.9](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-http@2.0.8...@equinor/fusion-framework-module-http@2.0.9) (2022-09-29)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-http
+
+
+
+
+
+## 2.0.8 (2022-09-27)
+
+
+### Bug Fixes
+
+* update registering of configuration ([20942ce](https://github.com/equinor/fusion-framework/commit/20942ce1c7a853ea3b55c031a242646e378db8c9))
+
+
+
+## 2.0.7 (2022-09-20)
+
+**Note:** Version bump only for package @equinor/fusion-framework-module-http
+
+
+
+
+
 ## [2.0.6](https://github.com/equinor/fusion-framework/compare/@equinor/fusion-framework-module-http@2.0.5...@equinor/fusion-framework-module-http@2.0.6) (2022-09-16)